refactor(wasm): split lib.rs (2703 → 266 lines) into api/ submodules
lib.rs was an encyclopedia — 170+ #[wasm_bindgen] functions covering
every domain: filters, paint tools, layer ops, adjustments, selection,
PSD I/O, the works. Moves each domain into its own file under src/api/,
leaving lib.rs as what it should have been: the Engine wrapper struct,
module declarations, and a small set of cross-cutting utilities
(compositing entry, coord transforms, CPU pixel helpers, color space,
export).

New modules (lines each):
  api/layer.rs       559   add/remove/update, upload (dense/sparse/
                           compressed/mask), duplicate/merge/flip/rotate/
                           scale, crop/resize canvas, clipboard copy/
                           cut/paste, floating selection, thumbnail read
  api/filter.rs      670   every GPU filter (Gaussian/box/unsharp, cel,
                           halftone, kaleidoscope, clouds, smoke, etc.)
                           + save/restore/clear preview
  api/paint.rs       272   brush / eraser / pencil / brush tip + dodge
                           burn / smudge / clone stamp
  api/psd.rs         257   exportPsd / parsePsd / getPsdLayerPixels /
                           getPsdLayerMask
  api/overlay.rs     168   grid, rulers, transform overlay, gradient
                           guide, path/lasso/crop/brush-cursor, plus
                           uploadLayerFromImageBitmap (canvas surface)
  api/adjustment.rs  171   exposure/contrast/highlights/shadows/whites/
                           blacks/vignette/sat/vibrance/curves/levels
                           + mask edit mode
  api/fill.rs        164   floodFill / applyFillToLayer /
                           readLayerPixelsForFill
  api/drawing.rs     139   gradient (linear/radial), shape render,
                           eyedropper sampleColor, upload path texture
  api/magnetic_lasso.rs 83 begin / snap / snapPoint / end
  api/selection.rs   75    mask set/clear + CPU selection builders

No JS-side changes needed: every #[wasm_bindgen(js_name = "...")]
keeps its export name, so src/engine-wasm/wasm-bridge.ts and its
consumers work unchanged. 171 exports before, 171 exports after — the
.d.ts diff is purely cosmetic (reordering by module).

Verified: npm run wasm:build:dev clean, npm run typecheck clean, 528
vitest tests pass.
